+++ datafusion/src/physical_plan/joins/mod.rs
@@ -15,6 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
+mod smj_utils;
 pub mod cross_join;
 pub mod hash_join;
 pub mod sort_merge_join;
