+++ datafusion/src/physical_plan/joins/sort_merge_join.rs
@@ -36,6 +36,7 @@ use crate::execution::runtime_env::RuntimeEnv;
 use crate::execution::runtime_env::RUNTIME_ENV;
 use crate::logical_plan::JoinType;
 use crate::physical_plan::expressions::{exprs_to_sort_columns, PhysicalSortExpr};
+use crate::physical_plan::joins::smj_utils::{join_arrays, range_len};
 use crate::physical_plan::joins::{
     build_join_schema, check_join_is_valid, column_indices_from_schema, comp_rows,
     equal_rows, ColumnIndex, JoinOn,
@@ -58,264 +59,6 @@ use tokio::sync::mpsc::{Receiver, Sender};
 
 type StringArray = Utf8Array<i32>;
 type LargeStringArray = Utf8Array<i64>;
-
-#[derive(Clone)]
-struct PartitionedRecordBatch {
-    batch: RecordBatch,
-    ranges: Vec<Range<usize>>,
-}
-
-impl PartitionedRecordBatch {
-    fn new(
-        batch: Option<RecordBatch>,
-        expr: &[PhysicalSortExpr],
-    ) -> Result<Option<Self>> {
-        match batch {
-            Some(batch) => {
-                let columns = exprs_to_sort_columns(&batch, expr)?;
-                let ranges = lexicographical_partition_ranges(
-                    &columns.iter().map(|x| x.into()).collect::<Vec<_>>(),
-                )?
-                .collect::<Vec<_>>();
-                Ok(Some(Self { batch, ranges }))
-            }
-            None => Ok(None),
-        }
-    }
-
-    #[inline]
-    fn is_last_range(&self, range: &Range<usize>) -> bool {
-        range.end == self.batch.num_rows()
-    }
-}
-
-struct StreamingBatch {
-    batch: Option<PartitionedRecordBatch>,
-    cur_row: usize,
-    cur_range: usize,
-    num_rows: usize,
-    num_ranges: usize,
-    is_new_key: bool,
-    on_column: Vec<Column>,
-    sort: Vec<PhysicalSortExpr>,
-}
-
-impl StreamingBatch {
-    fn new(on_column: Vec<Column>, sort: Vec<PhysicalSortExpr>) -> Self {
-        Self {
-            batch: None,
-            cur_row: 0,
-            cur_range: 0,
-            num_rows: 0,
-            num_ranges: 0,
-            is_new_key: true,
-            on_column,
-            sort,
-        }
-    }
-
-    fn rest_batch(&mut self, prb: Option<PartitionedRecordBatch>) {
-        self.batch = prb;
-        if let Some(prb) = &self.batch {
-            self.cur_row = 0;
-            self.cur_range = 0;
-            self.num_rows = prb.batch.num_rows();
-            self.num_ranges = prb.ranges.len();
-            self.is_new_key = true;
-        };
-    }
-
-    fn key_any_null(&self) -> bool {
-        match &self.batch {
-            None => return true,
-            Some(batch) => {
-                for c in self.on_column {
-                    let array = batch.batch.column(c.index());
-                    if array.is_null(self.cur_row) {
-                        return true;
-                    }
-                }
-                false
-            }
-        }
-    }
-
-    #[inline]
-    fn is_finished(&self) -> bool {
-        self.batch.is_none() || self.num_rows == self.cur_row + 1
-    }
-
-    #[inline]
-    fn is_last_key_in_batch(&self) -> bool {
-        self.batch.is_none() || self.num_ranges == self.cur_range + 1
-    }
-
-    fn advance(&mut self) {
-        self.cur_row += 1;
-        self.is_new_key = false;
-        if !self.is_last_key_in_batch() {
-            let ranges = self.batch.unwrap().ranges;
-            if self.cur_row == ranges[self.cur_range + 1].start {
-                self.cur_range += 1;
-                self.is_new_key = true;
-            }
-        } else {
-            self.batch = None;
-        }
-    }
-
-    fn advance_key(&mut self) {
-        let ranges = self.batch.unwrap().ranges;
-        self.cur_range += 1;
-        self.cur_row = ranges[self.cur_range].start;
-        self.is_new_key = true;
-    }
-}
-
-/// Holding ranges for same key over several bathes
-struct BufferedBatches {
-    /// batches that contains the current key
-    /// TODO: make this spillable as well for skew on join key at buffer side
-    batches: VecDeque<PartitionedRecordBatch>,
-    /// ranges in each PartitionedRecordBatch that contains the current key
-    ranges: VecDeque<Range<usize>>,
-    /// row index in first batch to the record that starts this batch
-    key_idx: Option<usize>,
-    /// total number of rows for the current key
-    row_num: usize,
-    /// hold found but not currently used batch, to continue iteration
-    next_key_batch: Option<PartitionedRecordBatch>,
-    /// Join on column
-    on_column: Vec<Column>,
-    sort: Vec<PhysicalSortExpr>,
-}
-
-#[inline]
-fn range_len(range: &Range<usize>) -> usize {
-    range.end - range.start
-}
-
-impl BufferedBatches {
-    fn new(on_column: Vec<Column>, sort: Vec<PhysicalSortExpr>) -> Self {
-        Self {
-            batches: VecDeque::new(),
-            ranges: VecDeque::new(),
-            key_idx: None,
-            row_num: 0,
-            next_key_batch: None,
-            on_column,
-            sort,
-        }
-    }
-
-    fn key_any_null(&self) -> bool {
-        match &self.key_idx {
-            None => return true,
-            Some(key_idx) => {
-                let first_batch = &self.batches[0].batch;
-                for c in self.on_column {
-                    let array = first_batch.column(c.index());
-                    if array.is_null(*key_idx) {
-                        return true;
-                    }
-                }
-                false
-            }
-        }
-    }
-
-    fn is_finished(&self) -> Result<bool> {
-        match self.key_idx {
-            None => Ok(true),
-            Some(_) => match (self.batches.back(), self.ranges.back()) {
-                (Some(batch), Some(range)) => Ok(batch.is_last_range(range)),
-                _ => Err(DataFusionError::Execution(format!(
-                    "Batches length {} not equal to ranges length {}",
-                    self.batches.len(),
-                    self.ranges.len()
-                ))),
-            },
-        }
-    }
-
-    /// Whether the running key ends at the current batch `prb`, true for continues, false for ends.
-    fn running_key(&mut self, prb: &PartitionedRecordBatch) -> Result<bool> {
-        let first_range = &prb.ranges[0];
-        let range_len = range_len(first_range);
-        let current_batch = &prb.batch;
-        let single_range = prb.ranges.len() == 1;
-
-        // compare the first record in batch with the current key pointed by key_idx
-        match self.key_idx {
-            None => {
-                self.batches.push_back(prb.clone());
-                self.ranges.push_back(first_range.clone());
-                self.key_idx = Some(0);
-                self.row_num += range_len;
-                Ok(single_range)
-            }
-            Some(key_idx) => {
-                let key_arrays = join_arrays(&self.batches[0].batch, &self.on_column);
-                let current_arrays = join_arrays(current_batch, &self.on_column);
-                let equal = equal_rows(key_idx, 0, &key_arrays, &current_arrays)?;
-                if equal {
-                    self.batches.push_back(prb.clone());
-                    self.ranges.push_back(first_range.clone());
-                    self.row_num += range_len;
-                    Ok(single_range)
-                } else {
-                    self.next_key_batch = Some(prb.clone());
-                    Ok(false) // running key ends
-                }
-            }
-        }
-    }
-
-    fn cleanup(&mut self) {
-        self.batches.drain(..);
-        self.ranges.drain(..);
-        self.next_key_batch = None;
-    }
-
-    fn reset_batch(&mut self, prb: &PartitionedRecordBatch) {
-        self.cleanup();
-        self.batches.push_back(prb.clone());
-        let first_range = &prb.ranges[0];
-        self.ranges.push_back(first_range.clone());
-        self.key_idx = Some(0);
-        self.row_num = range_len(first_range);
-    }
-
-    /// Advance the cursor to the next key seen by this buffer
-    fn advance_in_current_batch(&mut self) {
-        assert_eq!(self.batches.len(), self.ranges.len());
-        if self.batches.len() > 1 {
-            self.batches.drain(0..(self.batches.len() - 1));
-            self.ranges.drain(0..(self.batches.len() - 1));
-        }
-
-        if let Some(batch) = self.batches.pop_back() {
-            let tail_range = self.ranges.pop_back().unwrap();
-            self.batches.push_back(batch);
-            let next_range_idx = batch
-                .ranges
-                .iter()
-                .find_position(|x| x.start == tail_range.start)
-                .unwrap()
-                .0;
-            self.key_idx = Some(tail_range.end);
-            self.ranges.push_back(batch.ranges[next_range_idx].clone());
-            self.row_num = range_len(&batch.ranges[next_range_idx]);
-        }
-    }
-}
-
-fn join_arrays(rb: &RecordBatch, on_column: &Vec<Column>) -> Vec<ArrayRef> {
-    on_column
-        .iter()
-        .map(|c| rb.column(c.index()).clone())
-        .collect()
-}
 
 struct OutputBuffer {
     arrays: Vec<Box<dyn MutableArray>>,
@@ -1339,104 +1082,6 @@ impl SortMergeJoinDriver {
         }
     }
 
-    /// true for has next, false for ended
-    async fn advance_streamed(&mut self) -> Result<bool> {
-        if self.stream_batch.is_finished() {
-            self.get_stream_next().await?;
-            Ok(!self.stream_batch.is_finished())
-        } else {
-            self.stream_batch.advance();
-            Ok(true)
-        }
-    }
-
-    /// true for has next, false for ended
-    async fn advance_streamed_key(&mut self) -> Result<bool> {
-        if self.stream_batch.is_finished() || self.stream_batch.is_last_key_in_batch() {
-            self.get_stream_next().await?;
-            Ok(!self.stream_batch.is_finished())
-        } else {
-            self.stream_batch.advance_key();
-            Ok(true)
-        }
-    }
-
-    /// true for has next, false for ended
-    async fn advance_streamed_key_null_free(&mut self) -> Result<bool> {
-        let mut more_stream_keys = self.advance_streamed_key().await?;
-        loop {
-            if more_stream_keys && self.stream_batch.key_any_null() {
-                more_stream_keys = self.advance_streamed_key().await?;
-            } else {
-                break;
-            }
-        }
-        Ok(more_stream_keys)
-    }
-
-    async fn advance_buffered_key_null_free(&mut self) -> Result<bool> {
-        let mut more_buffered_keys = self.advance_buffered_key().await?;
-        loop {
-            if more_buffered_keys && self.buffered_batches.key_any_null() {
-                more_buffered_keys = self.advance_buffered_key().await?;
-            } else {
-                break;
-            }
-        }
-        Ok(more_buffered_keys)
-    }
-
-    /// true for has next, false for ended
-    async fn advance_buffered_key(&mut self) -> Result<bool> {
-        if self.buffered_batches.is_finished()? {
-            match &self.buffered_batches.next_key_batch {
-                None => {
-                    let batch = self.get_buffered_next().await?;
-                    match batch {
-                        None => return Ok(false),
-                        Some(batch) => {
-                            self.buffered_batches.reset_batch(&batch);
-                            if batch.ranges.len() == 1 {
-                                self.cumulate_same_keys().await?;
-                            }
-                        }
-                    }
-                }
-                Some(batch) => {
-                    self.buffered_batches.reset_batch(batch);
-                    if batch.ranges.len() == 1 {
-                        self.cumulate_same_keys().await?;
-                    }
-                }
-            }
-        } else {
-            self.buffered_batches.advance_in_current_batch();
-            if self.buffered_batches.batches[0]
-                .is_last_range(&self.buffered_batches.ranges[0])
-            {
-                self.cumulate_same_keys().await?;
-            }
-        }
-        Ok(false)
-    }
-
-    /// true for has next, false for buffer side ended
-    async fn cumulate_same_keys(&mut self) -> Result<bool> {
-        let batch = self.get_buffered_next().await?;
-        match batch {
-            None => Ok(false),
-            Some(batch) => {
-                let more_batches = self.buffered_batches.running_key(&batch)?;
-                if more_batches {
-                    self.cumulate_same_keys().await
-                } else {
-                    // reach end of current key, but the stream continues
-                    Ok(true)
-                }
-            }
-        }
-    }
-
     fn compare_stream_buffer(&self) -> Result<Ordering> {
         let stream_arrays =
             join_arrays(&self.stream_batch.batch.unwrap().batch, &self.on_streamed);
@@ -1448,18 +1093,6 @@ impl SortMergeJoinDriver {
             &stream_arrays,
             &buffer_arrays,
         )
-    }
-
-    async fn get_stream_next(&mut self) -> Result<()> {
-        let batch = self.streamed.next().await.transpose()?;
-        let prb = PartitionedRecordBatch::new(batch, &self.stream_batch.sort)?;
-        self.stream_batch.rest_batch(prb);
-        Ok(())
-    }
-
-    async fn get_buffered_next(&mut self) -> Result<Option<PartitionedRecordBatch>> {
-        let batch = self.buffered.next().await.transpose()?;
-        PartitionedRecordBatch::new(batch, &self.buffered_batches.sort)
     }
 }
